requirements.txt
asgiref==3.6.0
cftime==1.6.2
Django==4.2
djangorestframework==3.14.0
netCDF4==1.6.3
numpy==1.24.2
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
